backend/requirements.txt
# 体育教学辅助网站 - Python依赖包

# Web框架
fastapi>=0.128.0
uvicorn[standard]>=0.40.0

# 数据库
sqlalchemy>=2.0.45
alembic>=1.17.0
# psycopg2-binary  # 生产环境PostgreSQL驱动，开发环境可选择SQLite
# redis>=5.0.1     # 可选：Redis缓存

# 认证和安全
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# 数据处理
pandas>=2.3.3
openpyxl>=3.1.5
xlrd>=2.0.1

# 工具库
pydantic>=2.12.5
python-dotenv>=1.2.1
httpx>=0.25.2

# 日志和监控
structlog>=23.2.0
prometheus-client>=0.19.0

# 测试
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
factory-boy>=3.3.0

# 开发工具
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1